+++ src/ts/Ads/Ads.ts
@@ -340,6 +340,10 @@ export class Ads implements IAds {
             campaign.setTrackingUrls(trackingUrls);
         }
 
+        // First ad request within a game session can be made using recorded privacy information.
+        // If game method has changed since, it should be reset before e.g. showing consent dialog
+        this.resetOutdatedUserPrivacy();
+
         if (placement.getRealtimeData() && !this.isConsentShowRequired()) {
             this._core.Api.Sdk.logInfo('Unity Ads is requesting realtime fill for placement ' + placement.getId());
             const start = Date.now();
@@ -375,6 +379,16 @@ export class Ads implements IAds {
             this.showConsentIfNeeded(options).then(() => {
                 this.showAd(placement, campaign, options);
             });
+        }
+    }
+
+    private resetOutdatedUserPrivacy() {
+        const gamePrivacy = this.Config.getGamePrivacy();
+        const userPrivacy = this.Config.getUserPrivacy();
+        const gdprApplies = gamePrivacy.getMethod() !== PrivacyMethod.DEFAULT;
+        const methodHasChanged = userPrivacy.getMethod() !== gamePrivacy.getMethod();
+        if (gdprApplies && methodHasChanged) {
+            userPrivacy.clear();
         }
     }
 
@@ -507,7 +521,7 @@ export class Ads implements IAds {
 
         if(TestEnvironment.get('campaignId')) {
             CampaignManager.setCampaignId(TestEnvironment.get('campaignId'));
-            this.BackupCampaignManager.deleteBackupCampaigns();
+            this.BackupCampaignManager.setEnabled(false);
         }
 
         if(TestEnvironment.get('sessionId')) {